backend/pyproject.toml
[project]
name = "nexus-command"
version = "4.5.1"
description = "Enterprise AI Supervisor — LangGraph + Gemini 2.0 Flash + Guardian Node"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langfuse>=3.14.5",
    "langgraph>=1.0.9",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "slack-sdk>=3.40.1",
    "slowapi>=0.1.9",
    "structlog>=25.5.0",
    "supabase>=2.28.0",
    "tavily-python>=0.7.22",
    "uvicorn[standard]>=0.41.0",
    "faiss-cpu>=1.9.0",
    "beautifulsoup4>=4.12.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
